hphd demosaic: small speedup and own compilation unit, #5159
